@@ -1,3 +1,3 @@
-export { User } from './user';
-export { UserDto } from './user.dto';
+export * from './user';
+export * from './user.dto';
 export * from './ag-grid-date-column';
